Implement type check
