+++ golem-rib/src/call_type.rs
@@ -24,6 +24,23 @@ pub enum CallType {
     EnumConstructor(String),
 }
 
+impl CallType {
+    pub fn is_resource_method(&self) -> bool {
+        match self {
+            CallType::Function(parsed_fn_name) => {
+                dbg!(parsed_fn_name.to_static().clone());
+
+                parsed_fn_name
+                    .to_static()
+                    .function
+                    .resource_method_name()
+                    .is_some()
+            }
+            _ => false,
+        }
+    }
+}
+
 impl Display for CallType {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         match self {
