+++ golem-rib/src/type_checker/mod.rs
@@ -1,2 +1,6 @@
 pub(crate) use call_args::*;
+pub(crate) use pretty_analysed_type::*;
+pub(crate) use validate::*;
 mod call_args;
+mod pretty_analysed_type;
+mod validate;
